Claudepixelagents/requirements.txt
# AI Bot Architect - Продвинутая версия

# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn==0.24.0

# Telegram Bot
python-telegram-bot==20.7
python-dotenv==1.0.0
aiogram==3.4.1

# HTTP клиент для API запросов
httpx==0.25.2

# Веб-сервер для Hugging Face
Flask==2.3.3

# AI - Groq API
# (используем httpx напрямую)

# WhatsApp - Green-API
# (используем httpx напрямую)

# Google Sheets API
gspread==5.12.4
oauth2client==4.1.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Trigger.dev
trigger-sdk==0.3.0

# Веб-скрапинг и парсинг
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Работа с данными
pydantic==2.5.0
pandas==2.1.4
openpyxl==3.1.2

# Утилиты
pathlib2==2.3.7
typing-extensions==4.8.0
